Cronograma/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
gunicorn==23.0.0
cloudinary==1.41.0
MarkupSafe==3.0.2
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
certifi==2024.8.30
urllib3==2.2.3
six==1.16.0
psycopg2-binary==2.9.9
